backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
alembic>=1.13.0
strands-agents>=1.33.0
mcp>=1.0.0
openai>=1.30.0
python-socketio>=5.10.0,<6.0
python-engineio>=4.8.0,<5.0
aiohttp>=3.8.0,<3.10.0
celery[redis]>=5.3.0
redis>=5.0.0
sse-starlette>=1.6.0
tenacity>=8.2.0
structlog>=24.0.0
